Make Inline Text Title NOT required
When adding an Inline Text component to a course, the user was
being forced to add a title. For users who don't want a title,
they would often add a title, then check the "hide title and icon"
checkbox.  But when they would go back to modify the inline text
component, the title value would be clobbered, defeating the
purpose of requiring a title.

This removes that requirement.