Add yellow part highlighting to gradeallq page
This adds a hover-over highlighting on the gradeallq page to make
it easier for the teacher to see which score goes with which part of
the question.  Also refactors inline javascript into a separate file.
This fix renames some scoring boxes, so should not be deployed while
teacher may be working.

(cherry picked from commit ae385a2c89f835b118606809c45f2d0922a05013)

Cherry-picking this again to fix broken rubric functionality in prod.