Fixed: Incorrect course name on direct payment page in one case.
It looks like the course name is not always up to date in $sessiondata.
The "fix" for now is to just get the course name ourselves from the DB
using the course ID provided in the URL.

This is currently okay because we're not showing the paywall screens
when clicking around within an assessment. The course ID is NOT in the
URL in this case.

We only show paywall screens when clicking into an assessment from
another page, which should always contain the course ID in the URL.