Fixed: Continue to assessments URL is wrong on activation confirmation page.
WAYM-79: Activation Codes | OHM | Once a student has entered a valid
activation code, from the confirmation screen, they should be able to
proceed to the assessment.

From the activation code confirmation screen, when students click on this
link: https://ohm3.ludev.team/assessment/showtest.php

students are hitting an error page:

You are not authorized to view this page. If you are trying to reaccess
an assessment you've already started, access it from the course page