Bug fix: prevent viewing score in final score only mode
When only show final score is set, students were able to
view their score by doing one problem, viewing the summary,
then leaving the test and reentering it to continue working.
This fix changes it so that an attempt is burned up on any
unanswered questions when the summary is viewed in this mode.
A warning confirmation was added.

(cherry picked from commit 66111934d12d273d8d3adefdabbe69e7d6057db0)