Get next/prev in forum post to stay in newthreads/flaggedthreads
When a post was viewed from the new threads list, Next/Prev links were
going to next/prev thread in the forum, not the next/prev of the new thread
list.  This fixes that, and the same for flagged threads.

(cherry picked from commit 2a46b888765e46c317a4d2bdaf76408fd8895ccc)