For singlescore, only show one gradebox. Bug fix: grade for stu not showing after take
For a singlescore multipart question, including accounting questions, it's cleaner to have
a single scorebox show in the gradebook for teachers, and a single score to show for students.
Also, there was a bug preventing question scores from showing to students when the GB was set
to show them after the attempt was submitted.
IM-504

(cherry picked from commit ce46f11666572177de18b68df34b1fbc6d14e8d0)

Conflicts:
	assess2/vue/js/gbviewassess.js
	assess2/vue/js/gbviewassess.js.map