Bug fix: don't ignore latepass if startdate is before assessment start
This would happen when a latepass is used on a manual exception that was meant to open
the assessment early for someone.  In that case, the latepass date should be used instead
of the later assessment enddate.

(cherry picked from commit 8d60739b60a776a92969d0603518f1f6606e2dc7)