OHM-992: Force all Swagger UI requests to use http rather than https.
Swagger Lume collects the scheme of the request to determine whether to load assets via http or https. This poses a problem in hosted environments that are behind load balancers that convert https requests to http. A new helper function was added that inspects the env var called APP_LOAD_BALANCED. When this var is set as 'true' or not set at all then all requests are forced to https. When set to false or APP_ENV is 'local' or 'development' then requests are forced to http.